extract population simulation
